s01e01/requirements.txt
openai
instructor
python-dotenv
requests
